# Add multidimensional array and slice support