--- a/array.go
+++ b/array.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"encoding/binary"
 	"io"
+	"reflect"
 	"strconv"
 	"strings"
 	"unicode"
@@ -349,4 +350,25 @@ func QuoteArrayElementIfNeeded(src string) string {
 		return quoteArrayElement(src)
 	}
 	return src
+}
+
+func findDimensionsFromValue(value reflect.Value, dimensions []ArrayDimension, elementsLength int) ([]ArrayDimension, int, bool) {
+	switch value.Kind() {
+	case reflect.Array:
+		fallthrough
+	case reflect.Slice:
+		length := value.Len()
+		if 0 == elementsLength {
+			elementsLength = length
+		} else {
+			elementsLength *= length
+		}
+		dimensions = append(dimensions, ArrayDimension{Length: int32(length), LowerBound: 1})
+		for i := 0; i < length; i++ {
+			if d, l, ok := findDimensionsFromValue(value.Index(i), dimensions, elementsLength); ok {
+				return d, l, true
+			}
+		}
+	}
+	return dimensions, elementsLength, true
 }