--- a/bool_array.go
+++ b/bool_array.go
@@ -31,6 +31,7 @@ func (dst *BoolArray) Set(src interface{}) error {
 		}
 	}
 
+	// Attempt to match to select common types:
 	switch value := src.(type) {
 
 	case []bool:
@@ -84,6 +85,9 @@ func (dst *BoolArray) Set(src interface{}) error {
 			}
 		}
 	default:
+		// Fallback to reflection if an optimised match was not found.
+		// The reflection is necessary for arrays and multidimensional slices,
+		// but it comes with a 20-50% performance penalty for large arrays/slices
 		reflectedValue := reflect.ValueOf(src)
 		if !reflectedValue.IsValid() || reflectedValue.IsZero() {
 			*dst = BoolArray{Status: Null}
@@ -189,6 +193,7 @@ func (src *BoolArray) AssignTo(dst interface{}) error {
 	switch src.Status {
 	case Present:
 		if len(src.Dimensions) == 1 {
+			// Attempt to match to select common types:
 			switch v := dst.(type) {
 
 			case *[]bool:
@@ -212,6 +217,9 @@ func (src *BoolArray) AssignTo(dst interface{}) error {
 			}
 		}
 
+		// Fallback to reflection if an optimised match was not found.
+		// The reflection is necessary for arrays and multidimensional slices,
+		// but it comes with a 20-50% performance penalty for large arrays/slices
 		value := reflect.ValueOf(dst)
 		if value.Kind() == reflect.Ptr {
 			value = value.Elem()
@@ -251,9 +259,8 @@ func (src *BoolArray) assignToRecursive(value reflect.Value, index, dimension in
 		length := int(src.Dimensions[dimension].Length)
 		if reflect.Array == kind {
 			typ := value.Type()
-			typLen := typ.Len()
-			if typLen != length {
-				return 0, errors.Errorf("expected size %d array, but %s has size %d array", length, typ, typLen)
+			if typ.Len() != length {
+				return 0, errors.Errorf("expected size %d array, but %s has size %d array", length, typ, typ.Len())
 			}
 			value.Set(reflect.New(typ).Elem())
 		} else {