This PR adds support for multidimensional arrays and slices (https://www.postgresql.org/docs/current/arrays.html).

### Previously (/Motivation):
- Only single dimensional slices were supported.
- The single dimensional slice types were hard-coded.
- If a new type conversion was added to the array's element (e.g. `Int4`), this would not be automatically supported by the array itself (e.g. `Int4Array`) as it relies on a manual update.
- Arrays were not supported. This was likely due to the hard-coded slice types and it is not possible to hard-code all possible array sizes.

### New functionality:
- Multidimensional slices are now supported.
- Arrays are now also supported, including multidimensional arrays.
- Adds new test cases for multidimensional arrays and slices.
- New type conversions for an array are automatically supported when the array's element gains new type conversion support.

### Maintenance improvements:
- Removes hard-coded type conversions for arrays (e.g. `Int4Array`). Instead each array now relies on its element's (e.g. `Int4`) type conversion support.
- Simplifies `typed_array_gen.sh` generator script by removing the hard-coded single-dimensional types for arrays.

### This PR is fully backwards compatible:
- All previous test cases are unmodified and passed.
- Only small changes to the code:
  - Two functions were modified (`Set` and `AssignTo`) in `typed_array.go.erb` with the addition of two new unexported auxiliary functions for recursion.
  - One new unexported auxiliary function in `array.go` file
  - `go_array_types` parameter removal in `typed_array_gen.sh`
  - Additional tests for each (PG-)array to test the new multidimensional support for (Go-)arrays and slices.