@jackc 2020-08-07
I really like the increased flexibility and functionality. But there is a measurable performance hit (presumably from the reflection).

Given a benchmark of selecting a string slice with 1,000 elements the overall query time increased by ~20%.

Before:
```
jack@glados ~/dev/pgx ±master⚡ » got -run=NONE -bench=BenchmarkSelectArray -benchmem
goos: darwin
goarch: amd64
pkg: github.com/jackc/pgx/v4
BenchmarkSelectArray-16    	    4743	    252650 ns/op	  139218 B/op	    2013 allocs/op
PASS
ok  	github.com/jackc/pgx/v4	2.295s
```

After:
```
jack@glados ~/dev/pgx ±master⚡ » got -run=NONE -bench=BenchmarkSelectArray -benchmem
goos: darwin
goarch: amd64
pkg: github.com/jackc/pgx/v4
BenchmarkSelectArray-16    	    3900	    305444 ns/op	  139248 B/op	    2014 allocs/op
PASS
ok  	github.com/jackc/pgx/v4	2.339s 
```

This is given the benchmark:

```go
func BenchmarkSelectArray(b *testing.B) {
	conn := mustConnectString(b, os.Getenv("PGX_TEST_DATABASE"))
	defer closeConn(b, conn)

	b.ResetTimer()

	arraySize := 1000
	var input []string
	for i := 0; i < arraySize; i++ {
		input = append(input, "some elememt")
	}

	for i := 0; i < b.N; i++ {
		var output []string
		err := conn.QueryRow(context.Background(), "select $1::text[]", input).Scan(&output)
		if err != nil {
			b.Fatal(err)
		}
	}
}
```

I think it is reasonable for additional functionality to cost additional performance, but I would prefer the previous simple path not to take such a hit.

Would it make sense to to keep the previous specific type handling and only use reflection when that doesn't work? Or maybe just avoid reflection at the innermost part of the loop. I'm guessing most of the cost is this line `if err := dst.Elements[index].Set(value.Interface()); err != nil {`. Though, I haven't done any profiling so I could be totally wrong about where the performance hit is.